Added the distance_limit method to the Beans class
distance_limit tries to determine distance limits based on having no observed bursts. With a flux history and a set of GTIs, you generate burst trains at a range of distances and with a range of start times, and determine how many of them are consistent with the observations (i.e. all the bursts must fall in gaps)

Required modifications to the __init__ routine to overlook the burst data when it's not supplied.

Also made some changes to next_burst to make it fail more gracefully when the interpolation goes outside the extent of the data, and simplified generate_burst_train to cope better with failures of the burst search (in either forward or backward directions)

Added the plot_model method to display the results from runmodel; this is a bit clunky because the output from run_model can either be the packed model prediction array (for comparison with the data) OR a model dictionary with tagged time, fluence etc.

init method now also returns st, et as arrays, for gti_checking; previously these were passed as global variables, but I couldn't seem to actually access them

get_burst_train returns a dict that now has an iref element giving the reference burst index. This will help to keep track of more general cases where you might have multiple simulated bursts earlier as well as later