Fix unhashable slice error; tweaks to init funcs
The "unhashable slice" error seems to have arisen when calls to runmodel with
nonsensical parameters (Z<0) would mean that no burst times would be returned.
Attempted to fix this by calculating the prior first and ONLY if the
parameters are consistent with the prior, call the model.

Also simplified the init function, and removed a few redundant keyword
parameters from the Beans:__init__ function